pyproject.toml
[tool.poetry]
name = "youtube-videos-api"
version = "0.1.0"
description = ""
authors = ["Nitin Sanatan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.0.1"
elasticsearch = "^7.13.1"
python-dotenv = "^0.17.1"
google-api-python-client = "^2.70.0"
celery = "^5.2.7"
redis = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
